Fixing prevent critical defects in libprivilege-control and some log messages. Critical "Explicit null dereferenced" in function add_api_feture() in src/privilege-control.c fixed. Critical "Explicit null dereferenced" in function get_all_ids_internal() in src/access-db.c seems to be false warning.
[Issue#]        SSDWSSP-306
[Bug/Feature]   Fix prevent defects
[Cause]         Prevent server signalizes defects
[Solution]      N/A
[Verification]  Running tests and checking prevent output.

+++ src/privilege-control.c
@@ -2221,7 +2221,9 @@ API int add_api_feature(app_type_t app_type,
 	// remove both files in case of failure
 	if (ret != PC_OPERATION_SUCCESS) {
 		unlink(smack_file);
-		unlink(dac_file);
+		if (dac_file) {
+			unlink(dac_file);
+		}
 	}
 
 	return ret;
